pyproject.toml
[tool.poetry]
name = "certification-v2-2"
version = "0.1.0"
description = ""
authors = ["MaksRo52 <romancov_dzr@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
isort = "^6.0.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
